Improve structure of font module and exports
