+++ src/bevy/mod.rs
@@ -2,3 +2,7 @@ pub(crate) mod bbcode;
 pub(crate) mod conversion;
 pub(crate) mod font;
 pub(crate) mod plugin;
+
+pub use bbcode::{Bbcode, BbcodeBundle, BbcodeSettings};
+pub use font::*;
+pub use plugin::BbcodePlugin;
