+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub(crate) mod bbcode;
 pub(crate) mod bevy;
 
-pub use bevy::bbcode::{Bbcode, BbcodeBundle, BbcodeSettings};
-pub use bevy::plugin::BbcodePlugin;
+pub use bevy::*;
